Add validate check all , còn các quy chuẩn không có add , if else tới chết

+++ SCSE-BACKEND/Controllers/UserController.cs
@@ -9,10 +9,10 @@ using System.Net.Http;
 namespace SCSE_BACKEND.Controllers
 {
     [RoutePrefix("User")]
-    
+
     public class UserController : ApiController
     {
-       
+        string data = TokenManager.ValidateCheck();
         SCSE_DBEntities db = new SCSE_DBEntities();
         //------------------ Login -------------------------//
         [Route("Login")]
@@ -38,55 +38,68 @@ namespace SCSE_BACKEND.Controllers
         [HttpPost]
         public object AddOrEdiAccount(Account1 acc1)
         {
-            if (acc1.IDUser == 0)
+            string data = TokenManager.ValidateCheck();
+            if (data == "OK")
             {
-                Account acc = new Account
+                if (acc1.IDUser == 0)
                 {
-                    IDRole = acc1.IDRole,
-                    Email = acc1.Email,
-                    FullName = acc1.FullName,
-                    Username = acc1.Username,
-                    Password = Utils.GetMD5(acc1.Password),
-                    Phone = acc1.Phone,
-                    Image = acc1.Image,
-                    IDState = acc1.IDState,
-                    Sex = acc1.Sex,
-                    CreatedByDate = DateTime.Now
-                };
-                db.Accounts.Add(acc);
-                db.SaveChanges();
+                    Account acc = new Account
+                    {
+                        IDRole = acc1.IDRole,
+                        Email = acc1.Email,
+                        FullName = acc1.FullName,
+                        Username = acc1.Username,
+                        Password = Utils.GetMD5(acc1.Password),
+                        Phone = acc1.Phone,
+                        Image = acc1.Image,
+                        IDState = acc1.IDState,
+                        Sex = acc1.Sex,
+                        CreatedByDate = DateTime.Now
+                    };
+                    db.Accounts.Add(acc);
+                    db.SaveChanges();
+                    return new Response
+                    {
+                        Status = "Success",
+                        Message = "Data Success"
+                    };
+                }
+                else
+                {
+                    var obj = db.Accounts.Where(x => x.IDUser == acc1.IDUser).ToList().FirstOrDefault();
+                    if (obj.IDUser > 0)
+                    {
+                        obj.IDRole = acc1.IDRole;
+                        obj.Email = acc1.Email;
+                        obj.Username = acc1.Username;
+                        obj.IDState = acc1.IDState;
+                        obj.Phone = acc1.Phone;
+                        obj.FullName = acc1.FullName;
+                        obj.Image = acc1.Image;
+                        obj.Sex = acc1.Sex;
+                        db.SaveChanges();
+                        return new Response
+                        {
+                            Status = "Updated",
+                            Message = "Updated Successfully"
+                        };
+                    }
+                }
                 return new Response
                 {
-                    Status = "Success",
-                    Message = "Data Success"
+                    Status = "Error",
+                    Message = "Data not insert"
                 };
             }
             else
             {
-                var obj = db.Accounts.Where(x => x.IDUser == acc1.IDUser).ToList().FirstOrDefault();
-                if (obj.IDUser > 0)
+                return new Response
                 {
-                    obj.IDRole = acc1.IDRole;
-                    obj.Email = acc1.Email;
-                    obj.Username = acc1.Username;
-                    obj.IDState = acc1.IDState;
-                    obj.Phone = acc1.Phone;
-                    obj.FullName = acc1.FullName;
-                    obj.Image = acc1.Image;
-                    obj.Sex = acc1.Sex;
-                    db.SaveChanges();
-                    return new Response
-                    {
-                        Status = "Updated",
-                        Message = "Updated Successfully"
-                    };
-                }
+                    Status = "Error",
+                    Message = "Token Fail"
+                };
             }
-            return new Response
-            {
-                Status = "Error",
-                Message = "Data not insert"
-            };
+
         }
         // Mỹ yêu cầu
         [Route("EditPasswordAccount")]
@@ -114,22 +127,35 @@ namespace SCSE_BACKEND.Controllers
         [HttpPost]
         public object EditState(Account1 acc1)
         {
-            var obj = db.Accounts.Where(x => x.IDUser == acc1.IDUser).ToList().FirstOrDefault();
-            if (obj.IDUser > 0)
+            string data = TokenManager.ValidateCheck();
+            if (data == "OK")
             {
-                obj.IDState = acc1.IDState;
-                db.SaveChanges();
+
+                var obj = db.Accounts.Where(x => x.IDUser == acc1.IDUser).FirstOrDefault();
+                if (obj.IDUser > 0)
+                {
+                    obj.IDState = acc1.IDState;
+                    db.SaveChanges();
+                    return new Response
+                    {
+                        Status = "Updated",
+                        Message = "Updated Successfully"
+                    };
+                }
                 return new Response
                 {
-                    Status = "Updated",
-                    Message = "Updated Successfully"
+                    Status = "Error",
+                    Message = "Data not insert"
                 };
             }
-            return new Response
+            else
             {
-                Status = "Error",
-                Message = "Data not insert"
-            };
+                return new Response
+                {
+                    Status = "Error",
+                    Message = "Token Fail"
+                };
+            }
         }
         [Route("EditPersonalInformation")]
         [HttpPost]
@@ -161,39 +187,52 @@ namespace SCSE_BACKEND.Controllers
         [HttpGet]
         public object ShowAllAcount()
         {
-            
-                var a = (from acc in db.Accounts
-                         from quyen in db.Roles
-                         where quyen.IDRole == acc.IDRole
-                         select new
-                         {
-                             acc.IDUser,
-                             acc.FullName,
-                             quyen.RoleName,
-                             acc.Username,
-                             acc.Password,
-                             acc.Email,
-                             acc.CreatedByDate,
-                             acc.IDState,
-                             acc.Phone,
-                             acc.Image,
-                             acc.Sex
-                         }).ToList();
-                return a;
+
+            var a = (from acc in db.Accounts
+                     from quyen in db.Roles
+                     where quyen.IDRole == acc.IDRole
+                     select new
+                     {
+                         acc.IDUser,
+                         acc.FullName,
+                         quyen.RoleName,
+                         acc.Username,
+                         acc.Password,
+                         acc.Email,
+                         acc.CreatedByDate,
+                         acc.IDState,
+                         acc.Phone,
+                         acc.Image,
+                         acc.Sex
+                     }).ToList();
+            return a;
         }
         // Xóa tài khoản
         [Route("DeleteAccount")]
         [HttpDelete]
         public object DeleteAccount(int iduser)
         {
-            var obj = db.Accounts.Where(x => x.IDUser == iduser).ToList().FirstOrDefault();
-            db.Accounts.Remove(obj);
-            db.SaveChanges();
-            return new Response
+            string data = TokenManager.ValidateCheck();
+            if (data == "OK")
             {
-                Status = "Delete",
-                Message = "Delete Successfuly"
-            };
+
+                var obj = db.Accounts.Where(x => x.IDUser == iduser).ToList().FirstOrDefault();
+                db.Accounts.Remove(obj);
+                db.SaveChanges();
+                return new Response
+                {
+                    Status = "Delete",
+                    Message = "Delete Successfuly"
+                };
+            }
+            else
+            {
+                return new Response
+                {
+                    Status = "Error",
+                    Message = "Token Fail"
+                };
+            }
         }
         // GetByStateUser
         [Route("GetByStateUser")]
@@ -254,14 +293,26 @@ namespace SCSE_BACKEND.Controllers
         [HttpDelete]
         public object DeleteState(int IDState)
         {
-            var obj = db.States.Where(x => x.Id == IDState).ToList().FirstOrDefault();
-            db.States.Remove(obj);
-            db.SaveChanges();
-            return new Response
+            if (data == "OK")
             {
-                Status = "Delete",
-                Message = "Delete Successfuly"
-            };
+
+                var obj = db.States.Where(x => x.Id == IDState).ToList().FirstOrDefault();
+                db.States.Remove(obj);
+                db.SaveChanges();
+                return new Response
+                {
+                    Status = "Delete",
+                    Message = "Delete Successfuly"
+                };
+            }
+            else
+            {
+                return new Response
+                {
+                    Status = "Error",
+                    Message = "Token Fail"
+                };
+            }
         }
         [Route("GetByIdState")]
         [HttpGet]
@@ -331,6 +382,6 @@ namespace SCSE_BACKEND.Controllers
             var obj = db.Roles.Where(x => x.IDRole == idrole).ToList().FirstOrDefault();
             return obj;
         }
-        
+
     }
 }
